pyproject.toml
[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[project]
name = "fpdf2-textindex"
version = "0.1.0"
description = "Text Index for fpdf2"
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0-only"
license-files = ["LICENSE"]

authors = [
  { name = "Corvin Lasogga", email = "c.lasogga@gmail.com" }
]

keywords = ["pdf", "library", "markdown", "index", "textindex"]

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Operating System :: OS Independent",
  "Topic :: Printing",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Text Processing :: Markup",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Multimedia :: Graphics :: Presentation",
  "Typing :: Typed",
]

dependencies = [
    "fpdf2>=2.8.7",
    "typing-extensions>=4.15.0",
]

[dependency-groups]
dev = [
  "pre-commit>=4.5.1",
  "pylint>=4.0.5",
  "pytest>=8.4.2",
  "pytest-cov>=7.1.0",
  "pytest-md-report>=0.7.0",
  "ruff>=0.15.11",
]
docs = [
  "pdoc>=16.0.0",
]

[project.urls]
Homepage = "https://cola5.github.io/fpdf2-textindex/"
Documentation = "https://cola5.github.io/fpdf2-textindex/"
Code = "https://github.com/CoLa5/fpdf2-textindex"
"Issue tracker" = "https://github.com/CoLa5/fpdf2-textindex/issues"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true

[tool.coverage.run]
branch = true

[tool.coverage.report]
include = [
  "fpdf2_textindex/*.py",
  "fpdf2_textindex/**/*.py",
]
exclude_also = [
  # Don't complain about missing debug-only code:
  "def __repr__",

  # Don't complain if tests don't hit defensive assertion code:
  "raise AssertionError",
  "raise NotImplementedError",

  # Don't complain if non-runnable code isn't run:
  "if __name__ == .__main__.:",

  # Don't complain about abstract methods, they aren't run:
  "@(abc\\.)?abstractmethod",
]
ignore_errors = true
precision = 2

[tool.coverage.html]
directory = "public/coverage/"
title = "Coverage report of fpdf2_textindex"

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
addopts = "--cov-report=xml --cov=fpdf2_textindex"
log_cli = true
log_cli_level = "WARN"
md_report_color = "never"
md_report_exclude_outcomes = ["skipped"]
md_report_flavor = "github"

[tool.ruff]
line-length = 80
indent-width = 4
target-version = "py310"
fix = true

[tool.ruff.lint]
preview = true
select = [
  # flake8-annotations
  "ANN",
  # flake8-bugbear
  "B",
  # pydocstyle
  "D",
  # pydoclint
  "DOC",
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # flake8-future-annotations
  "FA",
  # flake8-fixme
  "FIX",
  # flake8-logging-format
  "G",
  # isort
  "I",
  # flake8-logging
  "LOG",
  # flake8-pytest-style
  "PT",
  # Ruff-specific rules
  "RUF",
  # flake8-simplify
  "SIM",
  # pyupgrade
  "UP",
]
extend-ignore = [
  # missing docstring in magic method
  "D105",
  # missing docstring in __init__
  "D107",
  # missing blank line after summary
  "D205",
]
fixable = ["ALL"]
unfixable = []

# Source: https://github.com/PyCQA/isort/blob/main/isort/profiles.py#L22
[tool.ruff.lint.isort]
case-sensitive = true
force-single-line = true
force-sort-within-sections = true
# lexicographical = true
single-line-exclusions = ["collections.abc", "typing", "typing_extensions"]
order-by-type = false
# group-by-package = true

[tool.ruff.lint.per-file-ignores]
"test/**" = ["D", "DOC"]  # Do not require docstrings in tests

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.typos.files]
extend-exclude = ["requirements-dev.txt"]

[tool.uv.build-backend]
module-name = "fpdf2_textindex"
module-root = ""
source_include = ["LICENSE", "README.md"]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
